requirements.txt
flask
gunicorn
python-docx
openpyxl